refactor: 受講生詳細画面をコンポーネントに分割しコードの見通しを良くしました
